Add explicit key get to save() for Django 1.4
Apparently, with Django 1.4, when save() is called and there is no existing _session_key on the object,
none is created. That means you end up with None as a session key. I have no idea what causes this to not work in 1.4, but this is
an attempt to fix it.